Move ingredients to component package
