+++ src/main/java/com/blamejared/jeitweaker/actions/ActionAddIngredient.java
@@ -1,11 +1,10 @@
 package com.blamejared.jeitweaker.actions;
 
 import com.blamejared.crafttweaker.api.actions.IUndoableAction;
-import com.blamejared.crafttweaker.api.brackets.CommandStringDisplayable;
 import com.blamejared.jeitweaker.state.JeiStateManager;
-import com.blamejared.jeitweaker.zen.JeiIngredient;
+import com.blamejared.jeitweaker.zen.component.JeiIngredient;
 
-public final class ActionAddIngredient<T extends CommandStringDisplayable, U> implements IUndoableAction {
+public final class ActionAddIngredient<T, U> implements IUndoableAction {
     
     private final JeiIngredient<T, U> ingredient;
