+++ src/main/java/com/blamejared/jeitweaker/actions/ActionHideIngredient.java
@@ -2,12 +2,11 @@ package com.blamejared.jeitweaker.actions;
 
 import com.blamejared.crafttweaker.api.CraftTweakerAPI;
 import com.blamejared.crafttweaker.api.actions.IUndoableAction;
-import com.blamejared.crafttweaker.api.brackets.CommandStringDisplayable;
 import com.blamejared.jeitweaker.state.JeiStateManager;
-import com.blamejared.jeitweaker.zen.JeiIngredient;
+import com.blamejared.jeitweaker.zen.component.JeiIngredient;
 import net.minecraftforge.fml.LogicalSide;
 
-public final class ActionHideIngredient<T extends CommandStringDisplayable, U> implements IUndoableAction {
+public final class ActionHideIngredient<T, U> implements IUndoableAction {
     
     private final JeiIngredient<T, U> ingredient;
