+++ src/main/java/com/blamejared/jeitweaker/zen/component/HackyJeiIngredientToMakeZenCodeHappy.java
@@ -1,19 +1,19 @@
-package com.blamejared.jeitweaker.zen;
+package com.blamejared.jeitweaker.zen.component;
 
 import com.blamejared.crafttweaker.api.annotations.ZenRegister;
 import com.blamejared.crafttweaker.api.brackets.CommandStringDisplayable;
 import com.blamejared.crafttweaker_annotations.annotations.Document;
 import org.openzen.zencode.java.ZenCodeType;
 
-@Document("mods/JEITweaker/JeiIngredient")
-@ZenCodeType.Name("mods.jei.JeiIngredient")
+@Document("mods/JEITweaker/component/JeiIngredient")
+@ZenCodeType.Name("mods.jei.component.JeiIngredient")
 @ZenRegister
 // TODO("Replace all usages of this with JeiIngredient once generic inference is better")
 public interface HackyJeiIngredientToMakeZenCodeHappy extends CommandStringDisplayable {
     
     @SuppressWarnings("unchecked")
-    default <T extends CommandStringDisplayable, U>  JeiIngredient<T, U> cast() {
+    default <T, U> JeiIngredient<T, U> cast() {
         
-        return (JeiIngredient<T, U>) (Object) this; // Mixin-cast
+        return (JeiIngredient<T, U>) this;
     }
 }
