+++ src/main/java/com/blamejared/jeitweaker/zen/component/IItemStackExpansions.java
@@ -1,4 +1,4 @@
-package com.blamejared.jeitweaker.zen;
+package com.blamejared.jeitweaker.zen.component;
 
 import com.blamejared.crafttweaker.api.annotations.ZenRegister;
 import com.blamejared.crafttweaker.api.item.IItemStack;
